add script template for each network
